feat: support --force option when archiving instance
