+++ frontend/src/components/Instance/index.ts
@@ -1,0 +1,3 @@
+import InstanceArchiveRestoreButton from "./InstanceArchiveRestoreButton.vue";
+
+export { InstanceArchiveRestoreButton };
