+++ frontend/src/locales/en-US.json
@@ -953,7 +953,8 @@
     "find-gcp-project-id-and-instance-id": "To find GCP Project ID and instance ID, see",
     "create-gcp-credentials": "To create Credentials, see",
     "used-for-testing-connection": "Used for testing connections only",
-    "all": "All instances"
+    "all": "All instances",
+    "force-archive-description": "Force to archive. All releated issues will be resolved, and all related sheets will be deleted."
   },
   "environment": {
     "id": "Environment ID",
