+++ frontend/src/locales/es-ES.json
@@ -946,7 +946,8 @@
     "find-gcp-project-id-and-instance-id": "Para encontrar el ID de proyecto y el ID de instancia de GCP, consulte",
     "create-gcp-credentials": "Para crear credenciales, consulte",
     "used-for-testing-connection": "Usado solo para probar conexiones",
-    "all": "Todas las instancias"
+    "all": "Todas las instancias",
+    "force-archive-description": "Forzar a archivar. \nTodos los problemas relacionados se resolverán y todas las hojas relacionadas se eliminarán."
   },
   "environment": {
     "id": "ID del Entorno",
